pyproject.toml
[project]
name = "gemfunc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "beautifulsoup4>=4.13.3",
    "bs4>=0.0.2",
    "colorama>=0.4.6",
    "docstring-parser>=0.16",
    "duckduckgo-search>=7.5.1",
    "google-genai>=1.3.0",
    "litellm>=1.62.1",
    "ollama>=0.4.7",
    "pillow>=11.1.0",
    "praw>=7.8.1",
    "psutil>=7.0.0",
    "pypdl>=1.5.2",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "thefuzz>=0.22.1",
    "wikipedia>=1.4.0",
]

[project.optional-dependencies]
windows = [
    "wmi>=1.5.1",
    "pywin32>=310; platform_system=='Windows'",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]